+++ src/outgoing/senders/mailboxes.py
@@ -1,5 +1,6 @@
 from abc import abstractmethod
 from email.message import EmailMessage
+import logging
 import mailbox
 from typing import List, Optional, TypeVar, Union
 from pydantic import PrivateAttr
@@ -7,6 +8,8 @@ from ..config import Path
 from ..util import OpenClosable
 
 T = TypeVar("T", bound="MailboxSender")
+
+log = logging.getLogger(__name__)
 
 
 class MailboxSender(OpenClosable):  # ABC inherited from OpenClosable
@@ -16,13 +19,19 @@ class MailboxSender(OpenClosable):  # ABC inherited from OpenClosable
     def _makebox(self) -> mailbox.Mailbox:
         ...
 
+    @abstractmethod
+    def _describe(self) -> str:
+        ...
+
     def open(self) -> None:
+        log.debug("Opening %s", self._describe())
         self._mbox = self._makebox()
         self._mbox.lock()
 
     def close(self) -> None:
         if self._mbox is None:
             raise ValueError("Mailbox is not open")
+        log.debug("Closing %s", self._describe())
         self._mbox.unlock()
         self._mbox.close()
         self._mbox = None
@@ -30,6 +39,11 @@ class MailboxSender(OpenClosable):  # ABC inherited from OpenClosable
     def send(self, msg: EmailMessage) -> None:
         with self:
             assert self._mbox is not None
+            log.info(
+                "Adding e-mail %r to %s",
+                msg.get("Subject", "<NO SUBJECT>"),
+                self._describe(),
+            )
             self._mbox.add(msg)
 
 
@@ -39,6 +53,9 @@ class MboxSender(MailboxSender):
 
     def _makebox(self) -> mailbox.mbox:
         return mailbox.mbox(self.path)
+
+    def _describe(self) -> str:
+        return f"mbox at {self.path}"
 
 
 class MaildirSender(MailboxSender):
@@ -54,6 +71,13 @@ class MaildirSender(MailboxSender):
             except mailbox.NoSuchMailboxError:
                 box = box.add_folder(self.folder)
         return box
+
+    def _describe(self) -> str:
+        if self.folder is None:
+            folder = "root folder"
+        else:
+            folder = f"folder {self.folder!r}"
+        return f"Maildir at {self.path}, {folder}"
 
 
 class MHSender(MailboxSender):
@@ -76,6 +100,15 @@ class MHSender(MailboxSender):
                     box = box.add_folder(f)
         return box
 
+    def _describe(self) -> str:
+        if self.folder is None:
+            folder = "root folder"
+        elif isinstance(self.folder, str):
+            folder = f"folder {self.folder!r}"
+        else:
+            folder = "folder " + "/".join(map(repr, self.folder))
+        return f"MH mailbox at {self.path}, {folder}"
+
 
 class MMDFSender(MailboxSender):
     configpath: Optional[Path] = None
@@ -84,6 +117,9 @@ class MMDFSender(MailboxSender):
     def _makebox(self) -> mailbox.MMDF:
         return mailbox.MMDF(self.path)
 
+    def _describe(self) -> str:
+        return f"MMDF mailbox at {self.path}"
+
 
 class BabylSender(MailboxSender):
     configpath: Optional[Path] = None
@@ -91,3 +127,6 @@ class BabylSender(MailboxSender):
 
     def _makebox(self) -> mailbox.Babyl:
         return mailbox.Babyl(self.path)
+
+    def _describe(self) -> str:
+        return f"Babyl mailbox at {self.path}"