+++ test/test_senders/test_mh.py
@@ -1,4 +1,5 @@
 from email.message import EmailMessage
+import logging
 from mailbox import MH
 from operator import itemgetter
 from pathlib import Path
@@ -33,8 +34,12 @@ def test_mh_construct(
 
 
 def test_mh_send_no_folder_new_path(
-    monkeypatch: pytest.MonkeyPatch, test_email1: EmailMessage, tmp_path: Path
+    caplog: pytest.LogCaptureFixture,
+    monkeypatch: pytest.MonkeyPatch,
+    test_email1: EmailMessage,
+    tmp_path: Path,
 ) -> None:
+    caplog.set_level(logging.DEBUG, logger="outgoing")
     monkeypatch.chdir(tmp_path)
     sender = from_dict(
         {
@@ -51,11 +56,33 @@ def test_mh_send_no_folder_new_path(
     msgs = list(inbox)
     assert len(msgs) == 1
     assert email2dict(test_email1) == email2dict(msgs[0])
+    assert caplog.record_tuples == [
+        (
+            "outgoing.senders.mailboxes",
+            logging.DEBUG,
+            f"Opening MH mailbox at {tmp_path/'inbox'}, root folder",
+        ),
+        (
+            "outgoing.senders.mailboxes",
+            logging.INFO,
+            f"Adding e-mail {test_email1['Subject']!r} to MH mailbox at"
+            f" {tmp_path/'inbox'}, root folder",
+        ),
+        (
+            "outgoing.senders.mailboxes",
+            logging.DEBUG,
+            f"Closing MH mailbox at {tmp_path/'inbox'}, root folder",
+        ),
+    ]
 
 
 def test_mh_send_folder_str_new_path(
-    monkeypatch: pytest.MonkeyPatch, test_email1: EmailMessage, tmp_path: Path
+    caplog: pytest.LogCaptureFixture,
+    monkeypatch: pytest.MonkeyPatch,
+    test_email1: EmailMessage,
+    tmp_path: Path,
 ) -> None:
+    caplog.set_level(logging.DEBUG, logger="outgoing")
     monkeypatch.chdir(tmp_path)
     sender = from_dict(
         {
@@ -73,11 +100,33 @@ def test_mh_send_folder_str_new_path(
     msgs = list(work)
     assert len(msgs) == 1
     assert email2dict(test_email1) == email2dict(msgs[0])
+    assert caplog.record_tuples == [
+        (
+            "outgoing.senders.mailboxes",
+            logging.DEBUG,
+            f"Opening MH mailbox at {tmp_path/'inbox'}, folder 'work'",
+        ),
+        (
+            "outgoing.senders.mailboxes",
+            logging.INFO,
+            f"Adding e-mail {test_email1['Subject']!r} to MH mailbox at"
+            f" {tmp_path/'inbox'}, folder 'work'",
+        ),
+        (
+            "outgoing.senders.mailboxes",
+            logging.DEBUG,
+            f"Closing MH mailbox at {tmp_path/'inbox'}, folder 'work'",
+        ),
+    ]
 
 
 def test_mh_send_folder_list_new_path(
-    monkeypatch: pytest.MonkeyPatch, test_email1: EmailMessage, tmp_path: Path
+    caplog: pytest.LogCaptureFixture,
+    monkeypatch: pytest.MonkeyPatch,
+    test_email1: EmailMessage,
+    tmp_path: Path,
 ) -> None:
+    caplog.set_level(logging.DEBUG, logger="outgoing")
     monkeypatch.chdir(tmp_path)
     sender = from_dict(
         {
@@ -97,6 +146,24 @@ def test_mh_send_folder_list_new_path(
     msgs = list(work)
     assert len(msgs) == 1
     assert email2dict(test_email1) == email2dict(msgs[0])
+    assert caplog.record_tuples == [
+        (
+            "outgoing.senders.mailboxes",
+            logging.DEBUG,
+            f"Opening MH mailbox at {tmp_path/'inbox'}, folder 'important'/'work'",
+        ),
+        (
+            "outgoing.senders.mailboxes",
+            logging.INFO,
+            f"Adding e-mail {test_email1['Subject']!r} to MH mailbox at"
+            f" {tmp_path/'inbox'}, folder 'important'/'work'",
+        ),
+        (
+            "outgoing.senders.mailboxes",
+            logging.DEBUG,
+            f"Closing MH mailbox at {tmp_path/'inbox'}, folder 'important'/'work'",
+        ),
+    ]
 
 
 def test_mh_send_no_folder_extant_path(